+++ src/syscall.c
@@ -5,17 +5,17 @@
 
 int io_setup(unsigned nr_events, aio_context_t *ctx_idp)
 {
-    return syscall(__NR_io_setup, nr_events, ctx_idp);
+    return (int)syscall(__NR_io_setup, nr_events, ctx_idp);
 }
 
 int io_destroy(aio_context_t ctx_id)
 {
-    return syscall(__NR_io_destroy, ctx_id);
+    return (int)syscall(__NR_io_destroy, ctx_id);
 }
 
 int io_submit(aio_context_t ctx_id, long nr, struct iocb **iocbpp)
 {
-    return syscall(__NR_io_submit, ctx_id, nr, iocbpp);
+    return (int)syscall(__NR_io_submit, ctx_id, nr, iocbpp);
 }
 
 int io_getevents(aio_context_t ctx_id,
@@ -24,7 +24,7 @@ int io_getevents(aio_context_t ctx_id,
                  struct io_event *events,
                  struct timespec *timeout)
 {
-    return syscall(__NR_io_getevents, ctx_id, min_nr, nr, events, timeout);
+    return (int)syscall(__NR_io_getevents, ctx_id, min_nr, nr, events, timeout);
 }
 
 #if HAVE_LINUX_IO_URING_H
@@ -33,12 +33,12 @@ int io_uring_register(int fd,
                       const void *arg,
                       unsigned int nr_args)
 {
-    return syscall(__NR_io_uring_register, fd, opcode, arg, nr_args);
+    return (int)syscall(__NR_io_uring_register, fd, opcode, arg, nr_args);
 }
 
 int io_uring_setup(unsigned int entries, struct io_uring_params *p)
 {
-    return syscall(__NR_io_uring_setup, entries, p);
+    return (int)syscall(__NR_io_uring_setup, entries, p);
 }
 
 int io_uring_enter(int fd,
@@ -47,7 +47,7 @@ int io_uring_enter(int fd,
                    unsigned int flags,
                    sigset_t *sig)
 {
-    return syscall(__NR_io_uring_enter, fd, to_submit, min_complete, flags, sig,
-                   _NSIG / 8);
+    return (int)syscall(__NR_io_uring_enter, fd, to_submit, min_complete, flags,
+                        sig, _NSIG / 8);
 }
 #endif
